Fixed a problem that was preventing searching for Twitter handles for races with more than 200 candidates.
