+++ import_export_ballotpedia/urls.py
@@ -12,4 +12,6 @@ urlpatterns = [
     url(r'^retrieve_candidates/$', views_admin.retrieve_candidates_from_api_view, name='retrieve_candidates'),
     url(r'^import_ballot_items_for_location/$', views_admin.import_ballot_items_for_location_view,
         name='import_ballot_items_for_location'),
+    url(r'^(?P<election_local_id>[0-9]+)/retrieve_distributed_ballotpedia_ballots/$',
+        views_admin.retrieve_distributed_ballotpedia_ballots_view, name='retrieve_distributed_ballotpedia_ballots'),
 ]
