+++ templates/candidate/candidate_edit.html
@@ -142,7 +142,9 @@ span.wrap_word { word-break: break-word; }
 </div>
 
 <div class="form-group">
-    <label for="ballotpedia_candidate_id_id" class="col-sm-3 control-label">Ballotpedia Candidate Id</label>
+    <label for="ballotpedia_candidate_id_id" class="col-sm-3 control-label">Ballotpedia Candidate Id
+        {% if candidate and candidate.ballotpedia_candidate_url %}(<a href="{{ candidate.ballotpedia_candidate_url }}" target="_blank">Go</a>){% endif %}
+    </label>
     <div class="col-sm-8">
         <input type="text" name="ballotpedia_candidate_id" id="ballotpedia_candidate_id_id" class="form-control"
                value="{% if candidate %}{{ candidate.ballotpedia_candidate_id|default_if_none:"" }}{% else %}{{ ballotpedia_candidate_id|default_if_none:"" }}{% endif %}" />
@@ -268,7 +270,9 @@ span.wrap_word { word-break: break-word; }
 </div>
 
 <div class="form-group">
-    <label for="ballotpedia_candidate_url_id" class="col-sm-3 control-label">Ballotpedia Candidate Page</label>
+    <label for="ballotpedia_candidate_url_id" class="col-sm-3 control-label">Ballotpedia Candidate Page
+        {% if candidate and candidate.ballotpedia_candidate_url %}(<a href="{{ candidate.ballotpedia_candidate_url }}" target="_blank">Go</a>){% endif %}
+    </label>
     <div class="col-sm-8">
         <input type="text" name="ballotpedia_candidate_url" id="ballotpedia_candidate_url_id" class="form-control"
                value="{% if candidate %}{{ candidate.ballotpedia_candidate_url|default_if_none:"" }}{% else %}{{ ballotpedia_candidate_url|default_if_none:"" }}{% endif %}" />
