+++ templates/candidate/candidate_list.html
@@ -86,8 +86,8 @@
         <li><a href="{% url 'import_export_twitter:scrape_social_media_for_candidates_in_one_election' %}?google_civic_election_id={{ google_civic_election_id }}" >
         Go to Candidate Sites and get Twitter Handles for this Election</a> (1-2 minutes)</li>
 
-        <li><a href="{% url 'twitter2:bulk_retrieve_possible_twitter_handles' %}?google_civic_election_id={{ google_civic_election_id }}&state_code={{ state_code }}&page={{ current_page_number }}" >
-        Retrieve Possible Twitter Handles</a> (10 at a time - about 30 seconds)</li>
+        <li><a href="{% url 'twitter2:bulk_retrieve_possible_twitter_handles' %}?google_civic_election_id={{ google_civic_election_id }}&state_code={{ state_code }}&page={{ current_page_number }}" target="_blank" >
+        Retrieve Possible Twitter Handles</a> (25 at a time - about 30 seconds - Open in New Window)</li>
 
         <li><a href="{% url 'google_custom_search:bulk_retrieve_possible_google_search_users' %}?google_civic_election_id={{ google_civic_election_id }}&state_code={{ state_code }}" >
         Retrieve Possible Candidates from Google</a> (20 at a time - 1-2 minutes)</li>
@@ -206,6 +206,9 @@
                         (<a href="{% url 'candidate:candidate_edit' candidate.id %}?google_civic_election_id={{ google_civic_election_id }}&state_code={{ state_code }}&show_all_twitter_search_results=1" target="_blank">
                             See all Options</a>)
                         {% endif %}
+                    </li></ul>
+                    {% elif candidate.no_twitter_possibilities_found %}
+                    <ul><li>no options found
                     </li></ul>
                     {% endif %}
                 {% endif %}
